+++ Player/Components/NetworkBehavior.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-public protocol NetworkLimitation {
+public protocol NetworkBehavior {
     /// Should specify the desired limit, in bits per second, of network bandwidth consumption allowed during playback or `nil` if no limit is required
-    var preferredMaxBitrate: Int64? { get set }
+    var preferredMaxBitrate: Int64? { get }
 }