+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Track selection for subtitles and audio
 * Preferred track option added to `TrackSelectable`. HLSNative now tries to set a prefered language on startup
 * Added seek methods taking a callback
+* Preferred bitrate limit
 
 #### Changes
 * Improved handling of bitrate and buffering events